Move ServerContext to clusteragent package.
