+++ cmd/cluster-agent/app/app.go
@@ -20,7 +20,6 @@ import (
 
 	"github.com/DataDog/datadog-agent/cmd/agent/common"
 	"github.com/DataDog/datadog-agent/cmd/cluster-agent/api"
-	"github.com/DataDog/datadog-agent/cmd/cluster-agent/api/types"
 	"github.com/DataDog/datadog-agent/cmd/cluster-agent/custommetrics"
 	"github.com/DataDog/datadog-agent/pkg/aggregator"
 	"github.com/DataDog/datadog-agent/pkg/clusteragent"
@@ -147,12 +146,6 @@ func start(cmd *cobra.Command, args []string) error {
 	aggregatorInstance := aggregator.InitAggregator(s, hostname)
 	aggregatorInstance.AddAgentStartupEvent(fmt.Sprintf("%s - Datadog Cluster Agent", version.DCAVersion))
 
-	clusterAgent, err := clusteragent.Run(aggregatorInstance.GetChannels())
-	if err != nil {
-		log.Errorf("Could not start the Cluster Agent Process.")
-
-	}
-
 	_, err = apiserver.GetAPIClient() // make sure we can connect to the apiserver
 	if err != nil {
 		log.Errorf("Could not connect to the apiserver: %v", err)
@@ -191,7 +184,7 @@ func start(cmd *cobra.Command, args []string) error {
 	}
 
 	// start the cmd HTTPS server
-	sc := types.ServerContext{
+	sc := clusteragent.ServerContext{
 		ClusterCheckHandler: clusterCheckHandler,
 	}
 	if err = api.StartServer(sc); err != nil {
@@ -206,7 +199,7 @@ func start(cmd *cobra.Command, args []string) error {
 	if config.Datadog.GetBool("external_metrics_provider.enabled") {
 		custommetrics.StopServer()
 	}
-	clusterAgent.Stop()
+	api.StopServer()
 	if stopCh != nil {
 		close(stopCh)
 	}
