+++ pkg/clusteragent/servercontext.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2018 Datadog, Inc.
 
-package types
+package clusteragent
 
 import "github.com/DataDog/datadog-agent/pkg/clusteragent/clusterchecks"
